pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.11",
    "discord-py>=2.5.2",
    "psutil>=7.0.0",
    "yt-dlp>=2025.6.9",
]
